fix(badge): renamed story and removed dependency to previwe-html styles
