+++ tegel/.storybook/preview-head.html
@@ -4,23 +4,4 @@
 
 <!-- Or you can load custom head-tag JavaScript: -->
 
-<style>
-  .badges-demo-box {
-    margin:5px;
-    height: 32px;
-    width: 32px;
-    position: relative;
-    background-color: #C4C4C4;
-  }
-  .badges-demo--default{
-    position: absolute;
-    left: 16px;
-    top: -5px;
-  }
-  .badges-demo--small{
-    position: absolute;
-    left: 26px;
-    top: -2px;
-  }
-</style>
