+++ tegel/src/components/badge/badges-component.tsx
