+++ tegel/src/components/badge/badges.scss
